requirements.txt
numpy==1.21.5
pandas==1.4.4
Pillow==9.4.0
requests==2.28.1
streamlit==1.15.1
lightgbm==3.3.5
altair<5

